Feature/rhea

+++ src/main/java/TEAM5/roomie/Repository/PostRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface PostRepository extends JpaRepository<Posts, Integer> {
+public interface PostRepository extends JpaRepository<Posts, Long> {
     List<Posts> findByTag(String groupBuyPosts);
 }
